pyproject.toml
[project]
name = "local-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "google-adk>=1.0.0",
    "google-generativeai>=0.8.5",
    "langchain-community>=0.3.22",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.45.1",
    "uvicorn>=0.34.2",
]
